+++ grails-app/controllers/org/pih/warehouse/order/ReceiveOrderWorkflowController.groovy
@@ -34,7 +34,7 @@ class ReceiveOrderWorkflowController {
 				// create a new shipment instance if we don't have one already
 				def orderCommand = new OrderCommand();
 				if (params.id) {
-					orderCommand = orderService.getOrder(params.id, session.user.id as int)
+					orderCommand = orderService.getOrder(params.id, session.user.id)
 				}
 				else {
 					flash.message = "${warehouse.message(code: 'default.not.found.message', args: [warehouse.message(code: 'order.label', default: 'Order'), params.id])}"
